+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## v3.3.6
 
+- If unrar is not available do not convert rar. Was making empty zips.
 - Support zip deflate64
 - Use unrar.cffi if available.